+++ scripts/eventcatalog-config-file-utils.js
@@ -1,38 +1,12 @@
 import { readFile, writeFile, rm } from 'node:fs/promises';
-import { existsSync } from 'fs';
+import { existsSync } from 'node:fs';
+import { copyFile } from 'node:fs/promises';
 import path from 'node:path';
 import { v4 as uuidV4 } from 'uuid';
 import { pathToFileURL } from 'url';
 
-// * Very strange behavior when importing ESM files from catalogs into core.
-//  * Core (node) does not know how to handle ESM files, so we have to try and convert them.
-//  *
-//  * This needs sorting out! Sorry if you are reading this, but it unblocked me for now!
-//  * @param {*} content
-//  * @returns
-//  */
-function convertESMtoCJS(content) {
-  // Replace import statements with require
-  content = content.replace(/import\s+([a-zA-Z0-9{},\s*]+)\s+from\s+['"]([^'"]+)['"];/g, (match, imports, modulePath) => {
-    return `const ${imports.trim()} = require('${modulePath}');`;
-  });
-
-  // Replace export default with module.exports
-  content = content.replace(/export\s+default\s+/g, 'module.exports = ');
-
-  // Replace named exports with module.exports
-  content = content.replace(/export\s+{([^}]+)}/g, (match, exports) => {
-    return `module.exports = {${exports.trim()}};`;
-  });
-
-  // Remove declarations of __filename and __dirname
-  content = content.replace(/^\s*(const|let|var)\s+__(filename|dirname)\s+=\s+.*;?\s*$/gm, '');
-
-  return content;
-}
-
 export async function cleanup(projectDirectory) {
-  const filePath = path.join(projectDirectory, 'eventcatalog.config.cjs');
+  const filePath = path.join(projectDirectory, 'eventcatalog.config.mjs');
   if (existsSync(filePath)) {
     await rm(filePath);
   }
@@ -40,24 +14,21 @@ export async function cleanup(projectDirectory) {
 
 export const getEventCatalogConfigFile = async (projectDirectory) => {
   try {
-    const rawFile = await readFile(path.join(projectDirectory, 'eventcatalog.config.js'), 'utf8');
+    let configFilePath = path.join(projectDirectory, 'eventcatalog.config.js');
 
-    // Have to conver the ESM to CJS...
-    const configAsCommonJS = convertESMtoCJS(rawFile);
+    const filePath = path.join(projectDirectory, 'package.json');
+    const packageJson = JSON.parse(await readFile(filePath, 'utf-8'));
 
-    await writeFile(path.join(projectDirectory, 'eventcatalog.config.cjs'), configAsCommonJS);
+    if (packageJson?.type !== 'module') {
+      await copyFile(configFilePath, path.join(projectDirectory, 'eventcatalog.config.mjs'));
+      configFilePath = path.join(projectDirectory, 'eventcatalog.config.mjs');
+    }
 
-    const configFilePath = path.join(projectDirectory, 'eventcatalog.config.cjs');
     const configFileURL = pathToFileURL(configFilePath).href;
-    const configAsCJS = await import(/* @vite-ignore */ configFileURL);
+    const config = await import(/* @vite-ignore */ configFileURL);
 
-    //  Clean up?
-    await writeFile(path.join(projectDirectory, 'eventcatalog.config.js'), rawFile);
-
-    await cleanup(projectDirectory);
-
-    return configAsCJS.default;
-  } catch (error) {
+    return config.default;
+  } finally {
     await cleanup(projectDirectory);
   }
 };
@@ -92,9 +63,7 @@ export const writeEventCatalogConfigFile = async (projectDirectory, newConfig) =
 
     // Write the updated content back to the file
     await writeFile(configFilePath, content);
-
-    await cleanup(projectDirectory);
-  } catch (error) {
+  } finally {
     await cleanup(projectDirectory);
   }
 };